requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cohere==5.20.0
colorama==0.4.6
comtypes==1.4.10
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.123.10
fastavro==1.12.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
frozenlist==1.8.0
fsspec==2025.3.2
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.3
langchain-classic==1.0.0
langchain-cohere==0.5.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==3.2.0
langchain-huggingface==1.1.0
langchain-openai==1.1.0
langchain-pinecone==0.2.13
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.4.56
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.9.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pillow==11.2.1
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.2.3
pypiwin32==223
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.98
pywin32==310
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
